Data and regression
The code for NDC has been changed such it includes a variable for orig. packager.
The code for consolidation has been clear - the stuff about regression is been moved to a separate file.
The code for regression has been consolidated in one file along with changes to the approach for data manipulation.